+++ orderApp/src/main/java/utez/edu/mx/orderapp/services/combos/ComboService.java
@@ -112,6 +112,11 @@ public class ComboService {
         Long comboId = Long.parseLong(comboDto.getComboId());
         Combo combo = comboRepository.findById(comboId)
                 .orElseThrow(() -> new EntityNotFoundException("Paquete no encontrado"));
+
+        if (!combo.getOrderCombos().isEmpty()) {
+            return new Response<>(null, true, 400, "No se puede eliminar el combo porque ha sido asociado con ordenes.");
+        }
+
         try {
             firebaseStorageService.deleteFileFromFirebase(combo.getComboImgUrl(), "combos-pics/");
         } catch (IOException e) {
